updated init files to accomodate renaming of files
