+++ mindtrace/jobs/mindtrace/jobs/__init__.py
@@ -1,23 +1,45 @@
-from mindtrace.jobs.types import BackendType, ExecutionStatus, Job, JobSchema, JobInput, JobOutput
-from mindtrace.jobs.utils import job_from_schema, ifnone
-from mindtrace.jobs.consumer import Consumer
-from mindtrace.jobs.orchestrator import Orchestrator
-from mindtrace.jobs.local.client import LocalClient
-from mindtrace.jobs.redis.client import RedisClient
-from mindtrace.jobs.rabbitmq.client import RabbitMQClient
+from mindtrace.jobs.mindtrace.jobs.consumer import Consumer
+from mindtrace.jobs.mindtrace.jobs.orchestrator import Orchestrator
+from mindtrace.jobs.mindtrace.jobs.types import BackendType, ExecutionStatus, Job, JobSchema, JobInput, JobOutput
+from mindtrace.jobs.mindtrace.jobs.utils import job_from_schema, ifnone
+from mindtrace.jobs.mindtrace.jobs.base.connection_base import BrokerConnectionBase
+from mindtrace.jobs.mindtrace.jobs.base.consumer_base import ConsumerBackendBase
+from mindtrace.jobs.mindtrace.jobs.base.orchestrator_backend import OrchestratorBackend
+from mindtrace.jobs.mindtrace.jobs.local.client import LocalClient
+from mindtrace.jobs.mindtrace.jobs.local.fifo_queue import LocalQueue
+from mindtrace.jobs.mindtrace.jobs.local.priority_queue import LocalPriorityQueue
+from mindtrace.jobs.mindtrace.jobs.local.stack import LocalStack
+from mindtrace.jobs.mindtrace.jobs.rabbitmq.client import RabbitMQClient
+from mindtrace.jobs.mindtrace.jobs.rabbitmq.connection import RabbitMQConnection
+from mindtrace.jobs.mindtrace.jobs.redis.client import RedisClient
+from mindtrace.jobs.mindtrace.jobs.redis.connection import RedisConnection
+from mindtrace.jobs.mindtrace.jobs.redis.fifo_queue import RedisQueue
+from mindtrace.jobs.mindtrace.jobs.redis.priority import RedisPriorityQueue
+from mindtrace.jobs.mindtrace.jobs.redis.stack import RedisStack
 
 __all__ = [
     'BackendType',
-    'Consumer',
     'ExecutionStatus', 
-    'ifnone',
     'Job', 
     'JobSchema', 
     'JobInput', 
     'JobOutput', 
     'job_from_schema', 
-    'LocalClient', 
+    'ifnone',
+    'Consumer',
     'Orchestrator', 
+    'BrokerConnectionBase',
+    'ConsumerBackendBase',
+    'OrchestratorBackend',
+    'LocalClient',
+    'LocalQueue',
+    'LocalPriorityQueue',
+    'LocalStack',
     'RabbitMQClient',
-    'RedisClient', 
+    'RabbitMQConnection',
+    'RedisClient',
+    'RedisConnection',
+    'RedisQueue',
+    'RedisPriorityQueue', 
+    'RedisStack'
 ]
