+++ mindtrace/jobs/mindtrace/jobs/rabbitmq/__init__.py
@@ -1,3 +1,4 @@
-from mindtrace.jobs.mindtrace.queue_management.rabbitmq.client import RabbitMQClient
-from mindtrace.jobs.mindtrace.queue_management.rabbitmq.connection import RabbitMQConnection
+from mindtrace.jobs.mindtrace.jobs.rabbitmq.client import RabbitMQClient
+from mindtrace.jobs.mindtrace.jobs.rabbitmq.connection import RabbitMQConnection
+
 __all__ = ["RabbitMQClient", "RabbitMQConnection"]
