+++ mindtrace/jobs/mindtrace/jobs/redis/__init__.py
@@ -1,3 +1,4 @@
-from mindtrace.jobs.mindtrace.queue_management.redis.client import RedisClient
-from mindtrace.jobs.mindtrace.queue_management.redis.connection import RedisConnection
+from mindtrace.jobs.mindtrace.jobs.redis.client import RedisClient
+from mindtrace.jobs.mindtrace.jobs.redis.connection import RedisConnection
+
 __all__ = ["RedisClient", "RedisConnection"]
